implement MarkProcessing and reject ops on processing resources

+++ pkg/storage/utils/decomposedfs/decomposedfs.go
@@ -928,6 +928,11 @@ func (fs *Decomposedfs) Move(ctx context.Context, oldRef, newRef *provider.Refer
 		return nil, err
 	}
 
+	// check processing on source
+	if oldNode.IsProcessing(ctx) {
+		return errtypes.ResourceProcessing(oldRef.String())
+	}
+
 	if err := fs.tp.Move(ctx, oldNode, newNode); err != nil {
 		return nil, err
 	}
@@ -1122,6 +1127,11 @@ func (fs *Decomposedfs) Delete(ctx context.Context, ref *provider.Reference) (dr
 		return nil, err
 	}
 
+	// check processing
+	if node.IsProcessing(ctx) {
+		return errtypes.ResourceProcessing(ref.String())
+	}
+
 	if err := fs.tp.Delete(ctx, node); err != nil {
 		return nil, err
 	}
@@ -1166,6 +1176,10 @@ func (fs *Decomposedfs) Download(ctx context.Context, ref *provider.Reference, o
 		return nil, nil, errtypes.NotFound(f)
 	}
 
+	// check processing
+	if n.IsProcessing(ctx) {
+		return nil, nil, errtypes.ResourceProcessing(ref.String())
+	}
 	ri, err := n.AsResourceInfo(ctx, rp, nil, []string{"size", "mimetype", "etag"}, true)
 	if err != nil {
 		return nil, nil, err
